Removing duplicate module
